feat: update predefined notifications from dialog to button
