+++ app/src/renderer/pages/navigator/application/metric-rules/index.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent, useMemo } from 'react';
 import Page from 'renderer/components/layout/Page';
 import { useNavigatorTree } from 'renderer/contexts/NavigatorTreeContext';
 import { Card } from '@mui/material';
-import { ApplicationRO } from '../../../../../common/generated_definitions';
+import { ApplicationRO } from 'common/generated_definitions';
 import ApplicationMetricRulesTable from './components/ApplicationMetricRulesTable';
 
 const ApplicationMetricRules: FunctionComponent = () => {
